tracing: split lightweight types out of tracing.hh into tracing_fwd.hh
Creates tracing/tracing_fwd.hh containing the small POD types
(trace_type, span_id, trace_state_props, trace_state_props_set,
trace_info, fmt::formatter<span_id>) that are needed by widely-included
headers but do not require the heavy tracing machinery (sharded<tracing>,
reactor, metrics, ...).

tracing.hh now includes tracing_fwd.hh and drops the duplicate type
definitions.  query/query-request.hh (included in ~329 TUs) switches to
include tracing_fwd.hh instead of tracing.hh, removing the full tracing
service from its transitive closure.

(cherry picked from commit f8980b31b5b51158da79a0a1f247a2d9c85ded4e,
originally part of scylladb/scylladb#29009)
